requirement.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-docx>=0.8.11
sentence-transformers>=0.4.1
faiss-cpu>=1.7.2
openai==1.82.0
pydantic>=1.8.2
python-multipart>=0.0.5
numpy>=1.21.0
pylatexenc==2.10